created a helper.go file in crontrollers,refactored error messages and updated the readme
helper.go will have all the helper functions required in
handlers/controllers
