+++ backend/controllers/helpers.go
@@ -1,0 +1,17 @@
+package controllers
+
+import "database/sql"
+
+// convert string to NullString
+func ToNullString(s string) sql.NullString {
+	if s == "" {
+		return sql.NullString{
+			String: s,
+			Valid:  false,
+		}
+	}
+	return sql.NullString{
+		String: s,
+		Valid:  true,
+	}
+}
