+++ backend/controllers/transactions.go
@@ -2,7 +2,6 @@ package controllers
 
 import (
 	"context"
-	"database/sql"
 	"net/http"
 	"time"
 
@@ -13,6 +12,7 @@ import (
 	"github.com/pelumitegbe/Personal-Finance-Tracker/models"
 )
 
+// handler for getting all the transactions
 func GetTransactions(db *database.Queries) gin.HandlerFunc {
 	return func(c *gin.Context) {
 		ctx, cancel := context.WithTimeout(context.Background(), 100*time.Second)
@@ -20,7 +20,7 @@ func GetTransactions(db *database.Queries) gin.HandlerFunc {
 
 		transactions, err := db.GetAllTransactions(ctx)
 		if err != nil {
-			c.JSON(http.StatusBadRequest, "couldn't get all the transactions")
+			c.JSON(http.StatusBadRequest, gin.H{"error": "Failed to retrieve transactions"})
 			return
 		}
 
@@ -28,34 +28,25 @@ func GetTransactions(db *database.Queries) gin.HandlerFunc {
 	}
 }
 
-// convert string to NullString
-func ToNullString(s string) sql.NullString {
-	if s == "" {
-		return sql.NullString{
-			String: s,
-			Valid:  false,
-		}
-	}
-	return sql.NullString{
-		String: s,
-		Valid:  true,
-	}
-}
-
 // function to add a transaction
 func AddTransaction(db *database.Queries) gin.HandlerFunc {
 	return func(c *gin.Context) {
+		// creating a context so that the request will timeout after certain time
 		ctx, cancel := context.WithTimeout(context.Background(), 100*time.Second)
-		var transactions models.Transaction
 		defer cancel()
 
+		// creating a transaction variable
+		var transactions models.Transaction
 		if err := c.BindJSON(&transactions); err != nil {
-			c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
+			c.JSON(http.StatusBadRequest, gin.H{"error": "Request body not valid"})
 			return
 		}
 
+		// changing string to  null string
 		category := ToNullString(transactions.Category)
 		description := ToNullString(transactions.Description)
+
+		// setting up our struct for pushing it into the database  with proper values
 		transactionData := database.AddTransactionsParams{
 			ID:              uuid.New(),
 			Amount:          transactions.Amount,
@@ -66,13 +57,19 @@ func AddTransaction(db *database.Queries) gin.HandlerFunc {
 			CreatedAt:       time.Now(),
 			UpdatedAt:       time.Now(),
 		}
+
+		// adding the transaction to the database
 		err := db.AddTransactions(ctx, transactionData)
 		if err != nil {
-			// c.JSON(http.StatusInternalServerError, gin.H{"error": "Not Created"})
+			c.JSON(
+				http.StatusInternalServerError,
+				gin.H{"error": "Couldn't create and store the transaction data"},
+			)
 
-			c.JSON(http.StatusInternalServerError, err.Error())
+			// c.JSON(http.StatusInternalServerError, err.Error())
 			return
 		}
-		c.JSON(http.StatusOK, "Successfully added the transaction to the database")
+		// succesfully returning once the transaction is added to the database
+		c.JSON(http.StatusCreated, "Successfully added the transaction to the database")
 	}
 }
